# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Prospix API
  description: |
    API multi-tenant da plataforma de prospecção inteligente Prospix.
    Primeiro tenant: Giovane Carrara (MetLife · SJRP).

    ## Autenticação
    Todas as rotas (exceto `/auth/*` e `/webhooks/*`) exigem JWT Bearer + header `X-Tenant-Id`.
    O `X-Tenant-Id` é validado contra o JWT — mismatch retorna 403.

    ## Multi-tenancy
    Todo dado é isolado por tenant via RLS no PostgreSQL. Um tenant nunca vê dados de outro.

    ## Rate limits
    - `/auth/*`: 10 req/min por IP
    - `/tenant/*` GET: 600 req/min por tenant
    - `/tenant/*` POST/PATCH: 60 req/min por tenant
  version: 1.0.0
  contact:
    name: Guilds
    email: gustavo.macedo@guilds.com.br
servers:
  - url: https://api.prospix.com.br/v1
    description: Produção
  - url: https://api-staging.prospix.com.br/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Local dev
...
security:
  - bearerAuth: []
    tenantHeader: []
paths:
...
  /tenant/integrations/google/callback:
    get:
      tags: [Settings]
      summary: Callback OAuth Google Calendar
      parameters:
        - { name: code, in: query, schema: { type: string } }
        - { name: state, in: query, schema: { type: string } }
        - { name: error, in: query, schema: { type: string } }
      responses:
        '302': { description: Redireciona para o app com status da integracao }
        '400': { description: Callback invalido ou state expirado }
...
  /tenant/integrations/whatsapp/connect:
    post:
      tags: [Settings]
      summary: Cria/verifica instancia WhatsApp e retorna QR Code
      x-prospix-response-shape: raw-object
      responses:
        '200':
          description: QR Code gerado
          content:
            application/json:
              schema:
                type: object
                properties:
                  instanceName: { type: string }
                  qrcode: { type: string }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { description: Falha externa Evolution API }
  /tenant/integrations/whatsapp/disconnect:
    post:
      tags: [Settings]
      summary: Desconecta sessao WhatsApp
      x-prospix-response-shape: raw-object
      responses:
        '200':
          description: Sessao desconectada
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string }
        '400': { description: Integracao nao configurada }
        '500': { description: Falha externa Evolution API }

  # ===========================================================================
  # NOTIFICATIONS
  # ===========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    tenantHeader:
      type: apiKey
      in: header
      name: X-Tenant-Id
  responses:
...
    NotFound:
      description: Recurso não encontrado neste tenant
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum: [UNAUTHENTICATED, UNAUTHORIZED, RESOURCE_NOT_FOUND, VALIDATION_ERROR, RATE_LIMITED, TENANT_QUOTA_EXCEEDED, EXTERNAL_SERVICE_DOWN, INTERNAL_ERROR]
            message: { type: string }
            details: { type: object }
            trace_id: { type: string }
        meta:
          type: object
          properties:
            request_id: { type: string }
            timestamp: { type: string, format: date-time }
...